4.21 | limit deleting blogs to user who added it

+++ part4/bloglist/controllers/blogs.js
@@ -46,8 +46,18 @@ blogsRouter.post('/', async (req, res, next) => {
 })
 
 blogsRouter.delete('/:id', async (req, res) => {
-  await Blog.findByIdAndRemove(req.params.id)
+  const decodedToken = jwt.verify(req.token, process.env.SECRET)
+  if(!decodedToken.id){
+    return res.status(401).json({ error: 'Token invalid' })
+  }
+  
+  const blog = await Blog.findById(req.params.id)
 
+  if(blog.user.toString() !== decodedToken.id.toString()){
+    res.status(401).json({ error: 'unauthorized operation' })
+  } 
+
+  await Blog.findByIdAndRemove(req.params.id)
   res.status(204).end()
 })
